The bug, per Daniel Zahn:
The bug, per Daniel Zahn:
1. When pixel scroll half a line, switch to line scroll a little, then
pixel scroll again the line is not offset correctly.
2. One time when pixel scrolling a little, then switching to line and
autoscrolling a little more, it cleared the current line.

His fix:
With praise like that how could I help but clear up the bug the very
same day. Seems to be working great now. For testing purposes the bug
would show when you pixel autoscrolled to halfway through a line,
switched to line autoscroll and scrolled a bit, then switched back and
pixel autscrolled. Now it doesn't happen. Turned out mostly it was a
cause of not setting the _locationChange flag in the linesDown and
linesUp function. Since Line Autoscroll is the only one who calls them
directly it wasn't a problem till now.